better plugin wrapper
Use `unsafeWindow`, thus no need to 'inject script in site context' anymore.
- wrapper is simpler now
- no need to cut it out in mobile app
  (but we keep that code for all legacy plugins)
- whole `scriptMetaStr` is added to plugin_info, to be able to analyse more properties
- it's easier to debug plugins now
  To set breakpoints switch to Web Developer Tools' Debugger/Sources tab
  and find related script in Tampermonkey's list of `userscript.html?id=<guid>`
- it's possible to use GM_* functions directly (not in mobile app, for now)
  - https://wiki.greasespot.net/Greasemonkey_Manual:API
  - https://www.tampermonkey.net/documentation.php
